+++ package.json
@@ -17,6 +17,7 @@
     "@emotion/styled": "^11.11.0",
     "@fontsource-variable/manrope": "^5.0.8",
     "@react-hook/window-size": "^3.1.1",
+    "@sanity/cli": "^3.23.1",
     "@sanity/client": "^6.4.9",
     "@sanity/image-url": "^1.0.2",
     "@sanity/vision": "^3.16.2",
