+++ sanity/package.json
@@ -15,6 +15,7 @@
     "sanity"
   ],
   "dependencies": {
+    "@sanity/cli": "^3.23.1",
     "@sanity/vision": "^3.15.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
