+++ sanity/schemas/index.ts
@@ -1,6 +1,8 @@
 import blockContent from './blockContent'
 import Headphones from './Headphones'
 import Speakers from './Speakers'
-import Earphones from "./Earphones"
+import Cars from './Cars'
 
-export const schemaTypes = [blockContent, Headphones, Speakers,Earphones]
+
+
+export const schemaTypes = [blockContent, Headphones, Speakers, Cars ]
