+++ src/components/pages/earphones/Earphones.tsx
@@ -12,7 +12,7 @@ import { AppContext } from "../../../context/AppContext";
 
 const Earphones: React.FC<IProduct> = () => {
   const { earphones } = useContext(AppContext);
-  console.log(earphones);
+ 
 
   return (
     <Flex
